# openapi.yaml
openapi: 3.1.0
info:
  version: 0.2.4
  title: Owncast APIs
  description: |-
    Internal
    The APIs required to make the Owncast web interface and internal functionality work. These APIs will never reach stability and will always change as Owncast features and functionality change. This means these endpoints are not good options to build your own tooling and features on top of. You can, however, of course can if that risk is acceptable to you.

    Admin
    The APIs used to configure the Owncast server, as well as power the web admin interface. These APIs are protected by the admin password, using HTTP Basic authentication. They should not be used by external tools or integrations as it would require you to give your admin password to a third party.
  contact:
    name: Gabe Kangas
    email: gabek@real-ity.com
    url: 'http://owncast.online'
  license:
    url: 'https://github.com/owncast/owncast/blob/develop/LICENSE'
    name: MIT
  x-logo:
    url: >-
...
servers:
  - url: 'http://localhost:8080/api'
    description: Localhost port 8080
paths:
...
  /chat:
    get:
      summary: Gets a list of chat messages
      operationId: GetChatMessages
      tags: ['Internal', 'Chat']
      parameters:
        - $ref: '#/components/parameters/AccessToken'
      responses:
        '200':
          description: List of chat messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessages'
        '401':
          $ref: '#/components/responses/401'
        '501':
          $ref: '#/components/responses/501'
  /chat/register:
    post:
      summary: Registers an anonymous chat user
      operationId: RegisterAnonymousChatUser
      tags: ['Internal', 'Chat']
      parameters:
        - in: header
          name: X-Forwarded-User
          schema:
            type: string
          required: false
      requestBody:
        description: The display name of the new user
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
      responses:
        '200':
          description: Anonymous user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnonymousUser'
        '400':
          $ref: '#/components/responses/400'
    options:
      operationId: RegisterAnonymousChatUserOptions
      x-internal: true
      tags: ['Objects', 'Chat']
      responses:
        '204':
          $ref: '#/components/responses/204Options'
...
  /yp:
    get:
      summary: Get the YP protocol data
      operationId: GetYPResponse
      tags: ['Internal']
      responses:
        '200':
          description: The current YP protocol configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YPDetails'
        '404':
          $ref: '#/components/responses/404'
...
  /integrations/chat:
    get:
      summary: Get chat history
      operationId: ExternalGetChatMessages
      tags: ['External', 'Chat']
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of chat messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessages'
        '401':
          $ref: '#/components/responses/401'
        '501':
          $ref: '#/components/responses/501'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: ExternalGetChatMessagesOptions
      x-internal: true
      tags: ['Objects', 'External', 'Admin', 'Chat']
      responses:
        '204':
          $ref: '#/components/responses/204'
...
components:
  schemas:
    Error:
      type: object
      description: Structure for an error response
      properties:
        error:
          type: string
    BaseAPIResponse:
      type: object
      description: Simple API response
      properties:
        message:
          type: string
        success:
          type: boolean
...
    AnonymousUser:
      type: object
      properties:
        id:
          type: string
        accessToken:
          type: string
        displayName:
          type: string
    User:
      type: object
      properties:
        createdAt:
          type: string
        disabledAt:
          type: string
        nameChangedAt:
          type: string
        id:
          type: string
        displayName:
          type: string
        previousNames:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        authProviders:
          type: array
          description: Friendly names of the external auth methods this user signed in with (e.g. IndieAuth, Fediverse, or a plugin slug). Absent for anonymous users.
          items:
            type: string
        displayColor:
          type: integer
        isBot:
          type: boolean
        authenticated:
          type: boolean
...
    Event:
      type: object
      properties:
        timestamp:
          type: string
        type:
          type: string
        id:
          type: string
    UserEvent:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        hiddenAt:
          type: string
        clientId:
          type: integer
    MessageEvent:
      type: object
      properties:
        body:
          type: string
    ChatMessages:
      type: array
      items:
        anyOf:
          - $ref: '#/components/schemas/UserMessage'
          - $ref: '#/components/schemas/SystemMessage'
          - $ref: '#/components/schemas/ActionMessage'
          - $ref: '#/components/schemas/FederatedAction'
    UserMessage:
      type: object
      allOf:
        - $ref: '#/components/schemas/Event'
        - $ref: '#/components/schemas/UserEvent'
        - $ref: '#/components/schemas/MessageEvent'
    SystemMessage:
      type: object
      allOf:
        - $ref: '#/components/schemas/Event'
        - $ref: '#/components/schemas/MessageEvent'
...
    ActionMessage:
      type: object
      allOf:
        - $ref: '#/components/schemas/Event'
        - $ref: '#/components/schemas/MessageEvent'
    FederatedAction:
      type: object
      allOf:
        - $ref: '#/components/schemas/Event'
        - $ref: '#/components/schemas/MessageEvent'
        - type: object
          properties:
            image:
              type: string
            link:
              type: string
            title:
              type: string
...
    SocialHandle:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
        icon:
          type: string
...
    YPDetails:
      type: object
      properties:
        lastConnectTime:
          type: string
        name:
          type: string
        description:
          type: string
        streamTitle:
          type: string
        logo:
          type: string
        tags:
          type: array
          items:
            type: string
        social:
          type: array
          items:
            $ref: '#/components/schemas/SocialHandle'
        viewerCount:
          type: integer
        overallMaxViewerCount:
          type: integer
        sessionMaxViewerCount:
          type: integer
        nsfw:
          type: boolean
        online:
          type: boolean
...
  headers:
    'Access-Control-Allow-Origin':
      schema:
        type: string
...
  responses:
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseAPIResponse'
    '204':
      description: No Content
    '204Options':
      description: No Content
      x-internal: true
      headers:
        'Access-Control-Allow-Origin':
          $ref: '#/components/headers/Access-Control-Allow-Origin'
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseAPIResponse'
    '401':
      description: Unauthorized
      content:
        text/plain:
          schema:
            type: string
...
    '404':
      description: Not found
...
    '501':
      description: Not Implemented
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    AccessToken:
      in: query
      name: accessToken
      schema:
        type: string
      required: true
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...